AI-Based-Smart-Inventory-Production/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
mysql-connector-python==8.2.0
PyMySQL==1.1.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
